Move graphql folder withing graph
