+++ pkg/controller/provider/web/graph/doc.go
@@ -1,4 +1,4 @@
-package graphql
+package graph
 
 import (
 	"github.com/konveyor/controller/pkg/inventory/container"
