+++ pkg/controller/provider/web/graph/resolver/vsphere/vm/resolver.go
@@ -7,7 +7,7 @@ import (
 	libmodel "github.com/konveyor/controller/pkg/inventory/model"
 	vspheremodel "github.com/konveyor/forklift-controller/pkg/controller/provider/model/vsphere"
 	graphmodel "github.com/konveyor/forklift-controller/pkg/controller/provider/web/graph/model"
-	"github.com/konveyor/forklift-controller/pkg/controller/provider/web/graphql/resolver"
+	"github.com/konveyor/forklift-controller/pkg/controller/provider/web/graph/resolver"
 )
 
 type Resolver struct {
